mirror the graphql crate in ffi
