+++ crates/iota-sdk-ffi/src/graphql/api/mod.rs
@@ -1,0 +1,17 @@
+// Copyright (c) 2026 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! GraphQL client FFI bindings.
+
+mod balance;
+mod checkpoints;
+mod coins;
+mod dry_run;
+mod dynamic_fields;
+mod epochs;
+mod events;
+mod iota_names;
+mod network;
+mod objects;
+mod package;
+mod transactions;
