+++ crates/iota-sdk-ffi/src/graphql/client.rs
@@ -1,0 +1,74 @@
+// Copyright (c) 2026 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use tokio::sync::RwLock;
+
+use crate::error::Result;
+
+/// The GraphQL client for interacting with the IOTA blockchain.
+#[derive(uniffi::Object)]
+pub struct GraphQLClient(pub(crate) RwLock<iota_sdk::graphql_client::Client>);
+
+impl GraphQLClient {
+    pub fn inner(&self) -> &RwLock<iota_sdk::graphql_client::Client> {
+        &self.0
+    }
+
+    pub fn into_inner(self) -> RwLock<iota_sdk::graphql_client::Client> {
+        self.0
+    }
+}
+
+#[uniffi::export(async_runtime = "tokio")]
+impl GraphQLClient {
+    // ===========================================================================
+    // Client Misc API
+    // ===========================================================================
+
+    /// Create a new GraphQL client with the provided server address.
+    #[uniffi::constructor]
+    pub fn new(server: String) -> Result<Self> {
+        Ok(Self(RwLock::new(iota_sdk::graphql_client::Client::new(
+            &server,
+        )?)))
+    }
+
+    /// Create a new GraphQL client connected to the `mainnet` GraphQL server:
+    /// {MAINNET_HOST}.
+    #[uniffi::constructor]
+    pub fn new_mainnet() -> Self {
+        Self(RwLock::new(iota_sdk::graphql_client::Client::new_mainnet()))
+    }
+
+    /// Create a new GraphQL client connected to the `testnet` GraphQL server:
+    /// {TESTNET_HOST}.
+    #[uniffi::constructor]
+    pub fn new_testnet() -> Self {
+        Self(RwLock::new(iota_sdk::graphql_client::Client::new_testnet()))
+    }
+
+    /// Create a new GraphQL client connected to the `devnet` GraphQL server:
+    /// {DEVNET_HOST}.
+    #[uniffi::constructor]
+    pub fn new_devnet() -> Self {
+        Self(RwLock::new(iota_sdk::graphql_client::Client::new_devnet()))
+    }
+
+    /// Create a new GraphQL client connected to the `localhost` GraphQL server:
+    /// {DEFAULT_LOCAL_HOST}.
+    #[uniffi::constructor]
+    pub fn new_localnet() -> Self {
+        Self(RwLock::new(iota_sdk::graphql_client::Client::new_localnet()))
+    }
+
+    /// Lazily fetch the max page size
+    pub async fn max_page_size(&self) -> Result<i32> {
+        Ok(self.0.read().await.max_page_size().await?)
+    }
+
+    /// Set the server address for the GraphQL GraphQL client. It should be a
+    /// valid URL with a host and optionally a port number.
+    pub async fn set_rpc_server(&self, server: String) -> Result<()> {
+        Ok(self.0.write().await.set_rpc_server(&server)?)
+    }
+}
