+++ crates/iota-sdk-ffi/src/graphql/faucet.rs
