+++ crates/iota-sdk-ffi/src/lib.rs
@@ -13,11 +13,9 @@ mod macros;
 
 pub mod crypto;
 pub mod error;
-pub mod faucet;
 pub mod graphql;
 pub mod transaction_builder;
 pub mod types;
-pub mod uniffi_helpers;
 
 uniffi::setup_scaffolding!();
